Add new workflow to update latest.json changes pushed to main (#84)
### Summary

Part of https://github.com/Shopify/theme-liquid-docs/issues/50

When changes are pushed to main, modify the latest.json file to indicate the repo has changed. 

We're moving this logic into a Github Action from the sync docs job to keep that job as small as possible and reducing the long term maintenance cost. In a Github Action, we can leverage the git cli vs having to maintain github ruby client code (which requires manually manipulating trees when creating commits)