Add support for theme blocks in sections
This commit adds support for theme blocks in sections.

Also does some refactoring for the intellisense docs of `type` in some cases.

For the allOf + `"if"` condition to work properly, we also need to add the
"required": ["type"] in there because otherwise the rule validates when the
property isn't present.